+++ Sources/KeyboardKit/Locales/LocaleFlagProvider.swift
@@ -15,30 +15,39 @@ import Foundation
  This protocol is implemented by both ``KeyboardLocale`` and
  `Locale`, where the implementations differ for some locales.
  */
-public protocol LocaleFlagProvider: LocaleProvider {
+public protocol LocaleFlagProvider {
 
     /**
-     The flag symbol for the ``LocaleProvider/locale``.
+     Get the locale flag symbol.
      */
     var flag: String { get }
 }
 
-public extension LocaleFlagProvider {
+@available(iOS 16, macOS 13, tvOS 16, watchOS 9, *)
+extension Locale: LocaleFlagProvider {}
+
+@available(iOS 16, macOS 13, tvOS 16, watchOS 9, *)
+public extension Locale {
 
     /**
-     The standard flag for the ``LocaleProvider/locale``.
+     Get the locale flag symbol.
+
+     This only works if the locale has a region identifier.
      */
-    var standardFlag: String {
-        countryFlag(for: locale.regionIdentifier ?? "")
+    var flag: String {
+        countryFlag(for: region?.identifier ?? "")
     }
 }
+
+@available(iOS 16, macOS 13, tvOS 16, watchOS 9, *)
+extension KeyboardLocale: LocaleFlagProvider {}
 
 private func countryFlag(for countryCode: String) -> String {
     let flagBase = UnicodeScalar("🇦").value - UnicodeScalar("A").value
     let flag = countryCode
         .uppercased()
         .unicodeScalars
-        .compactMap({ UnicodeScalar(flagBase + $0.value)?.description })
+        .compactMap { UnicodeScalar(flagBase + $0.value)?.description }
         .joined()
     return flag
 }